fix: type to int64

+++ internal/core/usecase/user_usecase.go
@@ -17,7 +17,7 @@ func (u *UserUsecase) CreateUser(account *domain.User) error {
 	return u.repo.CreateUser(account)
 }
 
-func (u *UserUsecase) GetUserByID(id int) (*domain.User, error) {
+func (u *UserUsecase) GetUserByID(id int64) (*domain.User, error) {
 	return u.repo.GetUserByID(id)
 }
 
@@ -25,6 +25,6 @@ func (u *UserUsecase) GetUserByEmail(email string) (*domain.User, error) {
 	return u.repo.GetUserByEmail(email)
 }
 
-func (u *UserUsecase) DeleteUser(id int) error {
+func (u *UserUsecase) DeleteUser(id int64) error {
 	return u.repo.DeleteUser(id)
 }
